+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update `@deltachat/stdio-rpc-server` and `deltachat/jsonrpc-client` to `v1.139.6`
 - Always show `msg.overrideSenderName` even when the message is sent by yourself
 - Update translations (2024-05-27)
+- Secure notifications on linux: escape html, like signal does
 
 ### Fixed
 - fix that map tab is highlighted when you are in the media/gallery tab #3867